Fix element panel properties form updating
