+++ js/main.ts
@@ -55,6 +55,7 @@ import "./interface/setup_settings"
 import "./edit_sessions"
 import "./validator"
 import "./outliner/outliner"
+import "./outliner/element_panel"
 import "./outliner/collections"
 import "./outliner/group"
 import "./outliner/mesh"
